Move Edge from coord sources
