+++ src/coord.hpp
@@ -1,7 +1,6 @@
 #ifndef PIET_I_COORD_HPP
 #define PIET_I_COORD_HPP
 
-#include <array>
 #include <tuple>
 #include <vector>
 #include "direction.hpp"
@@ -24,16 +23,6 @@ class Coordinates : public std::vector<Coord> {
   std::array<int, 4> range() const;
   bool includes(const Coord& coord) const;
   Coord find_out_of_range(const Coord& coord, Direction direction) const;
-};
-
-class Edges {
-  using BothChoose = std::array<Coord, 2>;
- public:
-  Edges(const Coordinates& coords);
-  Coord& edge(Direction direction, Choose choose);
-  const Coord& edge(Direction direction, Choose choose) const;
- private:
-  std::array<BothChoose, 4> edges_;
 };
 
 #endif  // PIET_I_COORD_HPP
